modify api

+++ message-io/src/test/java/net/easyrpc/message/io/test/Status.java
@@ -1,6 +1,6 @@
 package net.easyrpc.message.io.test;
 
 public class Status {
-    public int count;
+    public long count;
     public String message;
 }
